Made refactoring for some components and their hooks closes #67
