+++ src/builderGame/components/CreateTeamButton/CreateTeamButton.tsx
@@ -1,10 +1,22 @@
 import { Button, ButtonProps } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
-import { useCreateTeamButton } from './useCreateTeamButton';
+import { useNavigate } from 'react-router-dom';
+import { PATHS } from '../../../_shared/types';
+import { useCreateTeams, useSelectedPlayers } from '../../hook';
+import { generatesBalancedTeams } from '../../utils';
 
 export const CreateTeamButton: React.FC<ButtonProps> = props => {
-  const { createTeams } = useCreateTeamButton();
   const { t } = useTranslation();
+
+  const selectedPlayers = useSelectedPlayers();
+  const navigate = useNavigate();
+  const createMatchTeams = useCreateTeams();
+
+  function createTeams() {
+    const teams = generatesBalancedTeams(selectedPlayers);
+    createMatchTeams(teams);
+    navigate(`/${PATHS.PRE_MATCH}`);
+  }
 
   return (
     <Button size="lg" onClick={createTeams} {...props}>
