+++ src/builderGame/components/Team/Team.tsx
@@ -1,20 +1,30 @@
 import { HStack, Text } from '@chakra-ui/react';
-import { Teams } from '../../../_shared/types/general';
+import { useMemo } from 'react';
+import { PlayerCard } from '../../../_shared/components';
+import { TeamsName } from '../../../_shared/types/general';
+import { useFilteredPlayers } from '../../hook';
 import { TeamMaking } from '../../store/types';
-import { useTeam } from './useTeam';
 
 type Props = {
   teamMaking: TeamMaking;
-  team: Teams;
+  team: TeamsName;
 };
 
 export const Team: React.FC<Props> = ({ teamMaking, team }) => {
-  const { renderPlayers, teamName } = useTeam({ teamMaking, team });
+  const selectedPlayers = useFilteredPlayers(teamMaking.players);
+  const renderPlayers = useMemo(
+    () =>
+      selectedPlayers?.map(player => (
+        <PlayerCard {...player} team={team} key={player.id} size="small" />
+      )),
+    [selectedPlayers, team],
+  );
+
   return (
     <HStack gap="6">
       {renderPlayers}
       <Text as="h1" textStyle="team-name" color="brand.primary.regular">
-        {teamName}
+        {team}
       </Text>
     </HStack>
   );
