+++ src/builderGame/hook/useGetChosenPlayers.ts
@@ -1,0 +1,6 @@
+import { useSelector } from 'react-redux';
+import { getChosenPlayers } from '../store/selectors';
+
+export function useGetChosenPlayers() {
+  return useSelector(getChosenPlayers);
+}
