+++ src/builderGame/hook/useGetTeams.ts
@@ -1,0 +1,6 @@
+import { useSelector } from 'react-redux';
+import { getTeams } from '../store/selectors';
+
+export function useGetTeams() {
+  return useSelector(getTeams);
+}
