+++ src/builderGame/types/general.ts
@@ -1,0 +1,4 @@
+export type ColumnList = {
+  availablePlayers: string[];
+  selectedPlayers: string[];
+};
